application/backend/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
opensearch-py>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.10
python-dotenv>=1.0.0
